fix vimproc error when typescript is installed locally on Windows
- When typescript is installed locally to the project, tsuquyomi
attempts to invoke the command line:

  > node "%CD%\node_modules\typescript\bin\tsc

when getting the typescript version. The `\n` in `\node_modules` gets
interpreted as a newline for some reason.

The quick fix is to replace all instances of \\ with / in the command
line, as nodejs on Windows will read them as path separators just fine.