fix shortest import path edge case
For example, on a typescript project that uses the TypeORM node module,
when importing the symbol @Column.
This would actually navigate to a different buffer (./node_modules/typeorm/index.d.ts),
and even change it, instead of properly importing.

The reason for that is that an invalid vimgrep command would be built
and the code would jump to a different buffer (as the /j option in vimgrep wouldn't be interpreted)

For TypeORM @Column import, this would be the values for a:current_module_file:
'Column'
'columns/Column'
'decorator/columns/Column'
'[index]*'

The fix is to escape all '/' to generate a proper vimgrep command.